pyproject.toml
[project]
name = "aibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx==0.27.2",
    "jq>=1.11.0",
    "langchain-chatchat>=0.3.0.20240625",
    "PyYAML>=6.0",
    "pytest>=7.4.4",
]

[tool.pytest.ini_options]
# -s: 显示 print 输出
# -v: 显示详细的测试函数名
addopts = "-s -v"
asyncio_default_fixture_loop_scope = "function"
log_cli = true
log_cli_level = "DEBUG"
[tool.ruff.format]
# 设置缩进风格为 tab
indent-style = "space"